+++ backend/src/main/java/com/golden_pearl/backend/Models/User.java
@@ -24,14 +24,12 @@ public class User {
     private String accessKey;
     private Long contact;
     private Long joiningDate;
-    // private Integer investAmount;
-    // private Integer winAmount=0;
     private Integer withdrawAmount;
     private Integer balanceAmount;
     private Integer totalWin;
     @Builder.Default
     private boolean active = true;
-    private List<Short> loginTimeLines;
+    private List<Long> loginTimeLines;
     private List<Short> passwordResetTimeLines;
